api_sql/requirements.txt
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
Flask-Script==2.0.6
Jinja2==2.11.1
Mako==1.1.2
MarkupSafe==1.1.1
SQLAlchemy==1.3.15
Werkzeug==1.0.0
alembic==1.4.2
click==7.1.1
itsdangerous==1.1.0
python-dateutil==2.8.1
python-dotenv==0.12.0
python-editor==1.0.4
six==1.14.0
typing==3.7.4.1
